mp 分块 合并 下载已测试 删除 重命名 文件列表 回收站 真删除 测试完毕

+++ internel/service/file.go
@@ -250,7 +250,7 @@ func CompleteFileMpUpload(c *gin.Context) {
 	if num != req.ChunkCount {
 		c.JSON(400, response.NewRespone(errcode.FileStoreFail, "正在分块上传中", MultipartUploadCompleteRsp{
 			Completed: false,
-			Progress:  (num / req.ChunkCount) * 100,
+			Progress:  int((float32(num) / float32(req.ChunkCount)) * 100),
 		}))
 	} else {
 		//分块上传完毕 完成分块文件组装、数据库更新
